+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Migrate docs from terra-notification-dialog.
+
 ## 1.6.0 - (November 16, 2021)
 
 * Changed
