+++ packages/terra-framework-docs/package.json
@@ -48,6 +48,7 @@
     "terra-form-select": "^6.8.0",
     "terra-grid": "^6.0.0",
     "terra-modal-manager": "^6.58.0",
+    "terra-notification-dialog": "^4.22.0",
     "terra-section-header": "^2.0.0",
     "terra-slide-panel-manager": "^5.61.0"
   }
