+++ packages/terra-framework-docs/src/terra-dev-site/doc/notification-dialog/NotificationDialog.1.doc.mdx
@@ -3,17 +3,17 @@ import { Badge } from 'terra-notification-dialog/package.json?dev-site-package';
 import NotificationDialogPropsTable from 'terra-notification-dialog/src/NotificationDialog?dev-site-props-table';
 import ContentLayoutAsListPropsTable from 'terra-notification-dialog/src/_ContentLayoutAsList?dev-site-props-table';
 
-import HazardHighVariant from '../example/HazardHighVariant?dev-site-example';
-import HazardMediumVariant from '../example/HazardMediumVariant?dev-site-example';
-import HazardLowVariant from '../example/HazardLowVariant?dev-site-example';
-import ErrorVariant from '../example/ErrorVariant?dev-site-example';
-import CustomVariant from '../example/CustomVariant?dev-site-example';
+import HazardHighVariant from './example/HazardHighVariant?dev-site-example';
+import HazardMediumVariant from './example/HazardMediumVariant?dev-site-example';
+import HazardLowVariant from './example/HazardLowVariant?dev-site-example';
+import ErrorVariant from './example/ErrorVariant?dev-site-example';
+import CustomVariant from './example/CustomVariant?dev-site-example';
 
-import ContentNotificationDialog from '../example/ContentNotificationDialog?dev-site-example';
-import ActionOrderExample from '../example/ActionOrderExample?dev-site-example';
-import EmphasizedRejectExample from '../example/EmphasizedRejectExample?dev-site-example';
-import NotificationDialogInModalManager from '../example/NotificationDialogInModalManager?dev-site-example';
-import NotificationDialogWithLongText from '../example/NotificationDialogWithLongText?dev-site-example';
+import ContentNotificationDialog from './example/ContentNotificationDialog?dev-site-example';
+import ActionOrderExample from './example/ActionOrderExample?dev-site-example';
+import EmphasizedRejectExample from './example/EmphasizedRejectExample?dev-site-example';
+import NotificationDialogInModalManager from './example/NotificationDialogInModalManager?dev-site-example';
+import NotificationDialogWithLongText from './example/NotificationDialogWithLongText?dev-site-example';
 
 <Badge />
