+++ packages/terra-framework-docs/src/terra-dev-site/doc/notification-dialog/example/CustomVariant.jsx
