+++ packages/terra-framework-docs/src/terra-dev-site/test/notification-dialog/NotificationDialogWithNoMessage.test.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import Button from 'terra-button';
-import NotificationDialog from '../../../NotificationDialog';
+import NotificationDialog from 'terra-notification-dialog';
 
 const NotificationDialogNoMessage = () => {
   const [isOpen, setIsOpen] = useState(false);
