+++ packages/terra-framework-docs/src/terra-dev-site/test/notification-dialog/NotificationDialogWithNoTitle.test.jsx
@@ -1,19 +1,8 @@
 import React, { useState } from 'react';
 import Button from 'terra-button';
-import PropTypes from 'prop-types';
-import NotificationDialog from '../../../NotificationDialog';
+import NotificationDialog from 'terra-notification-dialog';
 
-const propTypes = {
-  variant: PropTypes.oneOf([
-    'hazard-high',
-    'hazard-medium',
-    'hazard-low',
-    'error',
-    'custom',
-  ]),
-};
-
-const NotificationDialogVariant = (props) => {
+const NotificationDialogNoTitle = () => {
   const [isOpen, setIsOpen] = useState(false);
 
   const handleOpenModal = () => {
@@ -24,21 +13,13 @@ const NotificationDialogVariant = (props) => {
     setIsOpen(false);
   };
 
-  const { variant, ...customProps } = props;
-
   return (
     <>
       {isOpen && (
         <NotificationDialog
-          {...customProps}
-          variant={variant}
-          dialogTitle="The title relates directly to the choices."
+          variant="error"
           startMessage="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
           acceptAction={{
-            text: 'Confirm',
-            onClick: handleCloseModal,
-          }}
-          rejectAction={{
             text: 'Close',
             onClick: handleCloseModal,
           }}
@@ -51,5 +32,4 @@ const NotificationDialogVariant = (props) => {
   );
 };
 
-NotificationDialogVariant.propTypes = propTypes;
-export default NotificationDialogVariant;
+export default NotificationDialogNoTitle;
