--- packages/terra-notification-dialog/src/terra-dev-site/doc/example/example-styles.module.scss
@@ -1,23 +1,0 @@
-@import '~terra-mixins/lib/Mixins';
-
-:local {
-  .example-wrapper {
-    height: 3rem;
-  }
-
-  .content-wrapper {
-    padding: 0.7rem;
-  }
-
-  .custom-notification-dialog-icon {
-    background: var(--terra-notification-dialog-example-custom-icon-background, inline-svg("<svg viewBox='0 0 49 49' xmlns='http://www.w3.org/2000/svg'><path fill='#78288C' d='M24.8 47.7l22.9-22.9c.4-.4.4-1.1 0-1.6L24.8.3c-.4-.4-1.1-.4-1.6 0L.3 23.2c-.4.4-.4 1.1 0 1.6l22.9 22.9c.4.4 1.2.4 1.6 0z' ></path></svg>"));
-  }
-
-  .orion-fusion-theme {
-    @include terra-inline-svg-var('--terra-notification-dialog-example-custom-icon-background', "<svg viewBox='0 0 49 49' xmlns='http://www.w3.org/2000/svg'><path fill='#78288C' d='M24.8 47.7l22.9-22.9c.4-.4.4-1.1 0-1.6L24.8.3c-.4-.4-1.1-.4-1.6 0L.3 23.2c-.4.4-.4 1.1 0 1.6l22.9 22.9c.4.4 1.2.4 1.6 0z' ></path></svg>");
-  }
-
-  .clinical-lowlight-theme {
-    @include terra-inline-svg-var('--terra-notification-dialog-example-custom-icon-background', "<svg viewBox='0 0 49 49' xmlns='http://www.w3.org/2000/svg'><path fill='#C270D7' d='M24.8 47.7l22.9-22.9c.4-.4.4-1.1 0-1.6L24.8.3c-.4-.4-1.1-.4-1.6 0L.3 23.2c-.4.4-.4 1.1 0 1.6l22.9 22.9c.4.4 1.2.4 1.6 0z' ></path></svg>");
-  }
-}
